% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abalone.analysis-package.R
\docType{package}
\name{abalone.analysis-package}
\alias{abalone.analysis}
\alias{abalone.analysis-package}
\title{abalone.analysis: Runs a Linear Regression Analysis on An Abalone Dataset}
\description{
Runs a simple linear regression analysis designed to predict the age of abalones based on their physical characteristics.
}
\keyword{internal}
