% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Create a linear regression model}
\usage{
build_model(
  func = "function for the model to analyze",
  training = "training data for the model"
)
}
\arguments{
\item{func}{The function for the model to analyze (not as a string)}

\item{training}{The training data for the function}
}
\value{
A fitted linear regression model
}
\description{
Creates a linear regression model using the training data, based on the diameter,
height, and shell weight of the abalone. Returns a fitted linear regression model
}
\examples{
build_model(func = age ~ diameter + height + shell_weight, training = abalone_training)
}
