% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean the \code{abalone_data} by reversing the scaling,
creating a new age target variable, (age = rings + 1.5)
removing the old target variable (rings)
and removing an unnecessary categorical variable (sex).}
\usage{
clean_data(data_frame = "the dataframe to be cleaned")
}
\arguments{
\item{data_frame}{Abalone dataframe with scaled values and the following columns:
sex, length, diameter, height, whole_weight, shucked_weight, shell_weight, rings}
}
\value{
a data_frame with reversed scaling, new target variable (age), removed columns (rings and sex)
}
\description{
Clean the \code{abalone_data} by reversing the scaling,
creating a new age target variable, (age = rings + 1.5)
removing the old target variable (rings)
and removing an unnecessary categorical variable (sex).
}
\examples{
x <- data.frame(length = c(1:10),
                diameter = c(1:10),
                height = c(1:10),
                whole_weight = c(1:10),
                shucked_weight = c(1:10),
                viscera_weight = c(1:10),
                shell_weight = c(1:10),
                sex = c("M", "F", "M", "F","M", "F","M", "F", "M", "F"),
                rings = c(1:10))
clean_data(data_frame = x)
}
