% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download and unpack the abalone data from the internet}
\usage{
download_data(url = "url to the abalone dataset")
}
\arguments{
\item{url}{The URL of the dataset to unpack}
}
\value{
An unpacked data file with the abalone metadata attached
}
\description{
Download and unpack the abalone data from the internet
}
\examples{
abalone_data <- download_data("https://archive.ics.uci.edu/static/public/1/abalone.zip")
}
