% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{split_data}
\alias{split_data}
\title{Creates the training and testing sets with a 70/30 train/test split using age for stratification.}
\usage{
split_data(clean_data)
}
\arguments{
\item{data_frame}{clean data that is ready to be split into training and testing (used clean_data)}
}
\value{
assigned training and testing sets to variables, returned as a list
}
\description{
Creates the training and testing sets with a 70/30 train/test split using age for stratification.
}
