% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Generate summary statistics table for numerical features}
\usage{
get_summary(dataset = "the dataset to clean")
}
\arguments{
\item{dataset}{A dataset or dataset extension (e.g. a tibble).}
}
\value{
A data frame with 6 columns.
The first column (named \code{variable}) lists the variables selected from
the original dataset (\code{age}, \code{shell_weight}, \code{diameter}, \code{height}).
The second column (named \code{mean}) contains the mean of the values for
each variable.
The third column (named \code{median}) contains the median of the values
for each variable.
The fourth column (named \code{variance}) contains the variance of the
values for each variable.
The fifth column (named \code{minimum}) contains the minimum value for
each variable.
The sixth column (named \code{maximum}) contains the maximum value for
each variable.
Each row corresponds to one of the variables selected from the
original dataset
}
\description{
Takes a dataset as input and calculates individual summary statistics
(mean, median, variance, minimum, and maximum) for all numeric columns
in a given dataset that has been reshaped from a wide format to a long
format
}
\examples{
get_summary(abalone_train)
}
